Return error if classifier doesn't exist before deleting

+++ cmd/ch360/commands/commands.go
@@ -1,6 +1,9 @@
 package commands
 
-import "github.com/CloudHub360/ch360.go/ch360"
+import (
+	"github.com/CloudHub360/ch360.go/ch360"
+	"errors"
+)
 
 type CreateClassifier struct {
 	client *ch360.ClassifiersClient
@@ -27,5 +30,23 @@ func NewDeleteClassifier(client *ch360.ClassifiersClient) *DeleteClassifier {
 }
 
 func (cmd *DeleteClassifier) Execute(classifierName string) error {
+	classifiers, err := cmd.client.GetAll()
+
+	if err != nil {
+		return err
+	}
+
+	if !classifierInList(classifierName, classifiers) {
+		return errors.New("There is no classifier named '" + classifierName + "'")
+	}
 	return cmd.client.DeleteClassifier(classifierName)
+}
+
+func classifierInList(classifierName string, list []ch360.Classifier) bool {
+	for _, b := range list {
+		if b.Name == classifierName {
+			return true
+		}
+	}
+	return false
 }
